move non-layout components out into their own folder
