+++ site/components/Hero.scss
@@ -1,4 +1,4 @@
-@import '../styles/config.scss';
+@import '../styles/config';
 
 .hero {
   display: flex;
